[SofaKernel] Move simulation_test into SofaSimulation.
