+++ SofaKernel/SofaSimulation/tests/simulation_test/scenes/PatchTestConstraint.scn
